refactor(review): remove rejected-design artifacts
- Move BudgetExceededError into a focused review error module.\n- Remove comments and tests that preserve discarded contract concepts.\n- Add agent rules for surgical changes, error placement, and comments.
