+++ src/review/README.md
@@ -8,11 +8,11 @@ A caller builds a `ReviewRequest`; an executor returns a `ReviewResult`.
 - `types.ts` — all review-domain interfaces (`ReviewTarget`, `ReviewRule`,
   `ReviewContext`, `ReviewBudget`, `ReviewFinding`, `ReviewScore`,
   `ReviewDiagnostic`, `ReviewResult`, `ReviewRequest`, `ReviewModelCall`).
-- `schemas.ts` — Zod schemas mirroring every external shape (boundary
-  validation). All schemas are strict; legacy scoring-mode, rubric, and
-  model-authored rule-override fields are rejected.
+- `schemas.ts` — strict Zod schemas mirroring every external shape (boundary
+  validation).
 - `budget.ts` — `DEFAULT_REVIEW_BUDGET`, `REVIEW_BUDGET_SCHEMA`,
-  `enforceBudget()`, and `BudgetExceededError` (extends `VectorlintError`).
+  and `enforceBudget()`.
+- `errors.ts` — review-domain errors, including `BudgetExceededError`.
 - `boundary.ts` — `buildScope()` / `isInScope()` enforcing the on-page boundary.
 - `executor.ts` — `ReviewExecutor` interface, `REVIEW_MODEL_CALLS`
   (`single | agent | auto`), `chooseModelCall()`.
@@ -33,9 +33,3 @@ A caller builds a `ReviewRequest`; an executor returns a `ReviewResult`.
 not how rules are scored. `single` sends target + rule in one structured call;
 `agent` gives the executor a target-scoped read-section capability; `auto`
 picks `single` for normal-sized inputs and `agent` for large ones.
-
-## Wiring status
-
-This module is additive and is **not** wired into the CLI yet. Future CLI
-wiring can emit `ReviewResult` through shared finding processing and implement
-executors behind `ReviewExecutor`.
