+++ src/review/errors.ts
@@ -1,0 +1,13 @@
+import { VectorlintError } from '../errors';
+import type { ReviewBudget } from './types';
+
+export class BudgetExceededError extends VectorlintError {
+  constructor(
+    message: string,
+    public readonly limit: keyof ReviewBudget,
+    public readonly actual: number,
+  ) {
+    super(message, 'BUDGET_EXCEEDED');
+    this.name = 'BudgetExceededError';
+  }
+}
